Moved hex_to_int and int_to_hex functions to imports/utils
